Move qscript optimizations to their own class.
