+++ sparkcore/src/main/scala/quasar/physical/sparkcore/fs/queryfile.scala
@@ -80,12 +80,13 @@ object queryfile {
       val lc: DiscoverPath.ListContents[F] =
         (adir: ADir) => EitherT(listContents(input, adir).liftM[PhaseResultT])
       val rewrite = new Rewrite[Fix]
+      val optimize = new Optimize[Fix]
       for {
         qs <- QueryFile.convertToQScriptRead[Fix, F, QScriptRead[Fix, ?]](lc)(lp)
         shifted <- Unirewrite[Fix, SparkQScriptCP, F](rewrite, lc).apply(qs)
 
         optQS = shifted.transHylo(
-                  rewrite.optimize(reflNT[SparkQScript]),
+                  optimize.optimize(reflNT[SparkQScript]),
                   Unicoalesce[Fix, SparkQScriptCP])
 
         _     <- EitherT(WriterT[Free[S, ?], PhaseResults, FileSystemError \/ Unit]((Vector(PhaseResult.tree("QScript (Spark)", optQS)), ().right[FileSystemError]).point[Free[S, ?]]))
